Fix: Checking website after deleting the file.

+++ uniforce/inc/scanner.php
@@ -138,17 +138,26 @@ function spbc_scanner_file_delete( $file_id ){
 
 					if($result){
 						
-						$response = Helper::http__request( CT_USP_URI ,array(),'dont_split_to_array');
-						if( Helper::http__request(CT_USP_URI, array(), 'get get_code') && ! Helper::search_page_errors($response) ){
-							// Deleting row from DB
-							$db->query('DELETE FROM scanner_files WHERE fast_hash = "'.$file_id.'"');
-						}else{
-							$output = array('error' =>'WEBSITE_RESPONSE_BAD');
-							$result = file_put_contents( $file_path, $remeber );
-							$output['error'] .= $result === false ? ' REVERT_FAILED' : ' REVERT_OK';
-						}
-						$output = array('success' => true);
-
+						$response = Helper::http__request(
+							CT_USP_URI,
+							array(),
+							'dont_split_to_array',
+							array( CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_2_0, )
+						);
+						
+						if( empty( $response['error'] ) ){
+							
+							if( Helper::http__request(CT_USP_URI, array(), 'get get_code') && ! Helper::search_page_errors($response) ){
+								// Deleting row from DB
+								$db->query('DELETE FROM scanner_files WHERE fast_hash = "'.$file_id.'"');
+							}else{
+								$output = array('error' =>'WEBSITE_RESPONSE_BAD');
+								$result = file_put_contents( $file_path, $remeber );
+								$output['error'] .= $result === false ? ' REVERT_FAILED' : ' REVERT_OK';
+							}
+							$output = array('success' => true);
+						}else
+							$output = $response;
 					}else
 						$output = array('error' =>'FILE_COULDNT_DELETE');
 				}else
